requirements.txt
Pillow>=9.0.0
mutagen>=1.45.1
